requirements.txt
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
toml==0.10.1
Werkzeug==1.0.1
xlrd==1.2.0
